Refactoring External user
Move and rename GoogleUser to an ExternalUser.
Google user was used for all types of user (oauth and LDAP for exemple) so I rename it on ExternalUser and move it to user module
