+++ kansha/user/comp.py
@@ -1,27 +1,24 @@
 # -*- coding:utf-8 -*-
-#--
+# --
 # Copyright (c) 2012-2014 Net-ng.
 # All rights reserved.
 #
 # This software is licensed under the BSD License, as described in
 # the file LICENSE.txt, which you should have received as part of
 # this distribution.
-#--
+# --
 
-import pkg_resources
 import peak.rules
 from datetime import datetime
 
 from nagare import presentation, component, i18n
 from nagare.security import common as security_common
-from nagare.i18n import _
 
 from . import usermanager
 from ..models import DataToken
 
 
 class User(security_common.User):
-
     """User component"""
 
     def __init__(self, username, *args, **kw):
@@ -177,7 +174,6 @@ def get_user_class(source):
 
 
 class PendingUser(object):
-
     """ Class for pending user of a board
 
     Store token id, to get the invited user
@@ -197,3 +193,23 @@ class PendingUser(object):
     @property
     def email(self):
         return self.data.username
+
+
+class ExternalUser(User):
+
+    def __init__(self, username, *args, **kw):
+        """Initialization
+
+        In:
+            - ``username`` -- the id of the user
+        """
+        super(User, self).__init__(username, 'passwd')
+        self.username = username
+        self._data = kw.get('data')
+
+
+@peak.rules.when(usermanager.get_user_class, """source != 'application'""")
+def get_user_class(source):
+    return ExternalUser
+
+
